requirements.txt
# Core dependencies
langgraph>=0.0.1
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Google Gemini
google-generativeai>=0.8.0

# Display and formatting
rich>=13.0.0
streamlit>=1.28.0

# Optional: For graph visualization
graphviz>=0.20.0
pygraphviz>=1.11
